+++ src/pic/simulator/parser/commands/Btfss.java
@@ -32,12 +32,12 @@ public class Btfss extends Command
 
 	@Override
 	public void execute(PicProcessor proc) {
-		byte val = proc.getMemoryControl().getAt(arg0);
+		short val = proc.getMemoryControl().getAt(arg0);
 		boolean bitIsSet = (val & (1<<arg1)) != 0;
 		
 		if(bitIsSet)
 		{
-			byte pcl = proc.getMemoryControl().getAt(SpecialFunctionRegister.PCL);
+			short pcl = proc.getMemoryControl().getAt(SpecialFunctionRegister.PCL);
 			pcl++;
 			proc.getMemoryControl().setAt(SpecialFunctionRegister.PCL, pcl);
 		}
